backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
supabase
python-dotenv
httpx
rembg
ultralytics
Pillow
google-generativeai
